backend/requirements.txt
# Core Framework
fastapi==0.119.0
uvicorn==0.37.0
pydantic==2.12.2
pydantic-settings==2.12.0
pydantic_core==2.41.4
starlette==0.48.0

# Authentication & Security
PyJWT==2.10.1

# HTTP Clients
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9

# Data Parsing
xmltodict==1.0.2

# Configuration
python-dotenv==1.1.1

# Database (for step 3)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Caching (for step 6)
redis==5.0.1

# OAuth (for step 5)
authlib==1.3.0

# Server & Utils
python-multipart==0.0.20
uvloop==0.21.0
watchfiles==1.1.1
websockets==15.0.1

# Other Dependencies
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
PyYAML==6.0.3
setuptools==80.9.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
